+++ tpm2_pytss/pyesys/utils.py
